+++ DDM3-Lite-Engine/DDM3-Lite/DataTypes/DescriptorObjects/TextureDescriptorObject.h
@@ -29,7 +29,7 @@ namespace DDM3
          
         void Clear();
 
-        int GetTextureAmount() const;
+        size_t GetTextureAmount() const;
 
         // Destructor
         virtual ~TextureDescriptorObject();
